pyproject.toml
[tool.poetry]
name = "goit-cs-hw-03"
version = "0.1.0"
description = ""
authors = ["AntonChubarov <achlp600@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.10"
faker = "^33.1.0"
pymongo = "^4.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
